[NFC][Vectorization] Refactor vector size inference out of the pass
Currently, the vector length inference lives inside the generic
vectorization pass. However, we need to infer vector lengths
when setting layouts for LLVMGPUVectorDistribute. This currently
happens prior to generic vectorization.

Therefore, this commit refactors inferSizesfromIR API into
codegen utils to be able to generally use it where its needed.
